--- a/crates/apollo-router/src/http_subgraph.rs
+++ b/crates/apollo-router/src/http_subgraph.rs
@@ -69,89 +69,44 @@ impl HttpSubgraphFetcher {
             .boxed()
     }
 
-    fn map_to_graphql(service_name: String, bytes_stream: BytesStream) -> graphql::ResponseStream {
-        // A `BytesStream` chunk doesn't always represent a `graphql::Response`.
-        // We need to accumulate the bytes until we have a deserialization that doesn't end up in EOF, and then yield a `graphql::Response`.
-        // `stream::unfold` allows us to do just that: keep some state around, and yield `Some((new_item, new_state))` until we're done (at which point we'll yield `None`).
-        //
-        // Here's the state we're keeping around:
-        //
-        // `bytes_stream` is the actual response we're trying to deserialize, and we'll repeatedly await.
-        //
-        // `current_payload_bytes` starts empty, and will accumulate until we have a full deserialization.
-        // It will then reset to an empty `BytesMut` and accumulate for the next payload.
-        //
-        // `service_name` is useful so we can correctly track and propagate errors.
-        //
-        // `is_primary` les the graphql::ResponseStream users know whether an error occured during a graphql::Response,
-        // or during a subsequent HTTP patch that appends an update.
-        stream::unfold(
-            (bytes_stream.fuse(), BytesMut::new(), service_name, true),
-            |(mut bytes_stream, mut current_payload_bytes, service_name, is_primary)| async move {
+    fn map_to_graphql(
+        service_name: String,
+        mut bytes_stream: BytesStream,
+    ) -> graphql::ResponseStream {
+        Box::pin(
+            async move {
+                let mut current_payload_bytes = BytesMut::new();
+                let is_primary = true;
+
                 while let Some(next_chunk) = bytes_stream.next().await {
                     match next_chunk {
                         Ok(bytes) => {
                             current_payload_bytes.extend(&bytes);
-                            match serde_json::from_slice::<graphql::Response>(
-                                &current_payload_bytes,
-                            ) {
-                                Err(e) if e.is_eof() => {
-                                    // Couldn't parse a full graphql::Response. This means the message is not complete yet.
-                                    continue;
-                                }
-                                Ok(response) => {
-                                    // Yield a graphql::Response
-                                    return Some((
-                                        response,
-                                        (bytes_stream, BytesMut::new(), service_name, false),
-                                    ));
-                                }
-                                Err(error) => {
-                                    // Yield a graphql::Response
-                                    return Some((
-                                        graphql::FetchError::SubrequestMalformedResponse {
-                                            service: service_name.clone(),
-                                            reason: error.to_string(),
-                                        }
-                                        .to_response(is_primary),
-                                        (bytes_stream, BytesMut::new(), service_name, false),
-                                    ));
-                                }
-                            }
                         }
                         Err(fetch_error) => {
-                            return Some((
-                                fetch_error.to_response(is_primary),
-                                (bytes_stream, BytesMut::new(), service_name, false),
-                            ));
+                            return fetch_error.to_response(is_primary);
                         }
                     }
                 }
 
-                // we're done with the `BytesStream`, yield a last response if any.
-                // since `bytes_stream` is fused, reentering the fold will lead us here,
-                // with an empty `current_payload_bytes` thus effectively yielding None,
-                // and exiting the fold
-                if !current_payload_bytes.is_empty() {
-                    let last_response =
-                        serde_json::from_slice::<graphql::Response>(&current_payload_bytes)
-                            .unwrap_or_else(|error| {
-                                graphql::FetchError::SubrequestMalformedResponse {
-                                    service: service_name.clone(),
-                                    reason: error.to_string(),
-                                }
-                                .to_response(is_primary)
-                            });
-                    Some((
-                        last_response,
-                        (bytes_stream, BytesMut::new(), service_name, false),
-                    ))
-                } else {
-                    None
+                match serde_json::from_slice::<graphql::Response>(&current_payload_bytes) {
+                    Ok(response) => {
+                        // Yield a graphql::Response
+
+                        response
+                    }
+                    Err(error) => {
+                        // Yield a graphql::Response
+                        graphql::FetchError::SubrequestMalformedResponse {
+                            service: service_name.clone(),
+                            reason: error.to_string(),
+                        }
+                        .to_response(is_primary)
+                    }
                 }
-            },
+            }
+            .into_stream(),
         )
-        .boxed()
     }
 }